requirements.txt
discord.py[voice]==2.4.0
yt-dlp==2024.11.4
python-dotenv==1.0.1
